Unmount component at the container node after render.
This fixes a memory leak by:

1) Providing the IsomorphicTag component, which allows you to render a
given html tag (like head and body) on the client but it will render a
`div` on the server, since React does not allow unmounting tags like
html, body and head.

2) autorender will get the innerHTML of the `div` tags containing the
children of the blacklisted tags mentioned above, and will generate a
valid HTML document to be sent to the client

The goal of this is to be able to call ReactDOM.unmountComponentAtNode
on the container element created during server side rendering so React
can free up its internal references, preventing a memory leak.

Closes #12